requirements.txt
streamlit
pandas
gspread
google-auth
google-auth-oauthlib
google-auth-httplib2
transformers
torch
sentencepiece
